Laison entre cours et user

+++ CoursServer/src/main/java/pfa/coursserver/repository/CoursRepository.java
@@ -5,6 +5,6 @@ import pfa.coursserver.entity.Cours;
 
 import java.util.List;
 
-public interface CoursRepository extends JpaRepository<Cours, Integer> {
+public interface CoursRepository extends JpaRepository<Cours, Long> {
 //    List<Cours> findByIdUser(Long userId);
 }
